# Improve hashset intersect with dictionary keycollection