@shubhamranjan 2020-08-26
There are some major faults in the new implemented behavior in terms of accuracy. One way to fix it would be to expose the dictionary's comparer through keycollection.